Update in-app-purchase to support packing via crosswalk-app-tools
Before the in-app-purchase was built via make_apk because long form permissions
(like com.android.vending.BILLING) cannot be added to crosswalk application
project via crosswalk-app-tools, but after the long form permisssions issue being
fixed, when we tried to build it via crosswalk-app-tools, it didn't work well.

Now we build a new apk via crosswalk-app-tools and upload this apk
to Google Play so this example can work well and does not depend on
the way of building apk.